rename NaturalLanguageNIO -> NaturalLanguage
