+++ Examples/Google/NaturalLanguage/RUNME
@@ -4,7 +4,7 @@
 # needed to build the example.
 #
 # Note that it requires updated protoc, protoc-gen-swift, and
-# protoc-gen-swiftgrpc binaries and assumes that protoc-gen-swift 
+# protoc-gen-swiftgrpc binaries and assumes that protoc-gen-swift
 # is installed in $HOME/local/bin.
 
 if [ ! -d "googleapis" ]; then
