--- Examples/Google/NaturalLanguageNIO/RUNME
@@ -1,28 +1,0 @@
-#!/bin/sh
-#
-# Use this script to regenerate the Protocol Buffer and gRPC files
-# needed to build the example.
-#
-# Note that it requires updated protoc, protoc-gen-swift, and
-# protoc-gen-swiftgrpc binaries and assumes that protoc-gen-swift
-# is installed in $HOME/local/bin.
-
-if [ ! -d "googleapis" ]; then
-  curl -L -O https://github.com/googleapis/googleapis/archive/master.zip
-  unzip master.zip
-  rm -f master.zip
-  mv googleapis-master googleapis
-fi
-
-protoc \
- 	google/cloud/language/v1/language_service.proto \
-	google/api/annotations.proto \
-	google/api/http.proto \
-	google/protobuf/descriptor.proto \
- 	-Igoogleapis \
-	-I../common/include \
-	--swift_out=googleapis \
-	--swiftgrpc_out=googleapis
-
-# move Swift files to the Sources directory
-find googleapis -name "*.swift" -exec mv {} Sources \;
